WoT Group C (C3): extract oversized-row/header measurement from baseTable.ts
The four post-render measurement methods (markOversizedRows, resetOversizedRows,
adjustColumnHeaderHeights, syncOversizedColumnHeadersWithFrozenOverlays) move to
axisSizing/oversizedRows.ts as a mixin, co-located with the axis-sizing slice (kept by design as the
'is a second pass needed?' signal). No subclass overrides; all callers (drawCycle, overlays) use the
Table type, so the declaration-merge mixin keeps them working. #deps reads route through the get deps()
getter; byte-identical logic. Gated green: test:types + eslint + test:walkontable 749/0/2 + core e2e
colHeader/nestedHeaders/autoRowSize/mergeCells/getRowHeight/renderSizeProbe 837/0.
